move util function to utils.py
